requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
python-dotenv==1.0.0
aiofiles==23.2.1